NOMERGE: add MPU configuration to panic output
This updates the panic! output to include the MPU Configuration:

    MPU Configuration:
     0: 0x00030000-0x00030200 Sub ✘'s 0x00 S ✘ X ✔ B ✘ C ✘ TEX 0 AP 6
     1: 0x20004000-0x20006000 Sub ✘'s 0x00 S ✘ X ✔ B ✘ C ✘ TEX 0 AP 3
     2: Disabled.
     3: Disabled.
     4: Disabled.
     5: Disabled.
     6: Disabled.
     7: Disabled.

As implemented, this is a big bucket of hacks. We don't currently save
the `Region`s created by `setup_mpu`, which means that the panic would
need to recreate them, but panic isn't passed a chip MPU instance, so
it can't.

I don't want to trample over the work that @alevy is doing to refactor
the MPU stuff, but this is immediately useful for debugging, hence
throwing this out into the world.

@alevy: Grab the printing bits from `debug_mpu_hack` and add it to
arch/cortex-m4 whenever that's ready