Fixs MPU bug that creates non-power-of-two regions
To find subregion sizes that align with the base address, we were
simply doing `start % len`. This only results in a power-of-two
subregion size when `start` only has 1`s in bits more significant
than than those in `len` (intuitively, when the non-power-of-two bits in
`start` are divisible by `len`).

Instead, we find the largest power-of-two that `start` is aligned to
(divisible by), by simply finding the lowest 1-bit in `start` and using
it as the highest (and only) bit in the subregion size. The function
later validates that this is otherwise a valid subregion (>= 32 and
results in a valid region size).