Add driver for SAM4L's USBC and generic USB lib
This is still experimental, and provides just enough support so that
the SAM4L can be enumerated as a device (at least on Linux).

In the boot code for imix, the kernel attaches to the USB bus and
registers handlers to respond to host requests.

This commit includes:

- The beginnings of a generic interface to USB controllers
  in kernel/src/hil
- A driver (implementing the hil-controller interface) for the
  SAM4L's USBC in chips/sam4l
- A library in capsules/src/usb.rs that provides datatypes and
  serialization routines for speaking USB
- A client (implementing the hil-client interface) in
  capsules/src/usb_simple.rs that does all the platform-independent
  logic necessary for enumeration, like responding to host requests for
  various descriptors
- Some unsafe support code in kernel/src/common, most of which
  is not pretty and can probably be removed before mainlining

There is no user-space interface yet.