Avoid requiring an explicit size in static_init
Sacrifices up to a word of memory in order to avoid requiring the caller
to explicitly pass in the size of the type. When we finally get
`size_of` or other intrinsics at compile-time, we'll be able to get rid
of that overhead as well.

For backwards compatibility accepts an optional size argument which is
simply ignored.