travis: ; -> || exit
popd was returning success even if the script failed. It doesn't matter
that we won't be in the right directory as travis will quit once one
command fails anyway.

n.b. You can't chain the popd with &&, as that won't do what you want either,
you need the explicit exit to maximize robustness
http://mywiki.wooledge.org/BashFAQ/105