SAM4L keep track of GPIO interrupt count for PM
Instead of enabling/disabling the GPIO clock, which would require
enabling it on each operation, we just special case GPIO for power
management and keep track of how many interrupts are enabled through a
shared, atomic variable. This works because we _only_ care about
interrupts for determining whether it's OK to go into various sleep
modes.