Fixes two bugs in the SPI code introduced in 7217345d
- real_rate: We corrected the check for whether the division was exact
from `clock % rate` to `clock % real_rate`, but that introduced a new
bug, because then it actually tries to increment it when it is already
at the slowest rate (which ends up setting scbr to the invalid 0). The
solution is to only get a slower baud rate if we are not at the slowest
rate already.
- transfers_in_progress: the hunk that set this to 0 was accidentally
deleted when we squashed/reordered some commits so that our code
introduces SPI slave and then reverts the board initialization code.