Make compression return the number of bytes consumed from the header
Change the signature of compress to take in the entire IPv6 datagram
and return both the number of bytes consumed and bytes written. This
will be needed when actually sending the packet because we will need
to determine the remaining length of the payload.

Consequently, change the IPv6 encapsulation and UDP compression code to
only consume the header, leaving the payload alone.