Refactored power manager interface
Added a new method of specifying external oscillator and system clock in
order to allow adding slow startup which is necessary for sleep on
signpost.

Tried to make the system clock specification more rigid. It has always
been assumming a 16 MHz external oscillator and allowed you to set a 48
MHz PLL based on it, or use it directly. Now those options are explict.

Also moved functionality into bpm, bscif, scif, and flashcalw as
appropriate while I was in here since a lot of functionality was handled
in PM.