Modify `read_write_bytes` DMA ordering
After fixing all the clock issues, I found that the signpost
`signpost_controller_app` still wasn't working and eventually traced the
problem down to not receiving a DMA read callback in the SPI over USART
implementation. Debugging with GPIO pins showed that the RX DMA transfer
was not started until after several bytes had already been written over
the wire, which seems bad. Combining the maps together and then starting
the RX transfer first solved the issue.

It is entirely unclear to me why this would be an issue now when it has
worked for months...

After further testing though, we have switched back to original SPI DMA ordering

It appears that changing the map stuff was the most important, because
the `signpost_controller_app` still works. Changing the ordering and
starting the read first hear led to an extra byte in the receive buffer
which current drivers and applications did not account for.