process: fix containers that impl Drop.
The code was assigning Default::default() into allocated memory, which
implies dropping the value that previously lived there. But the "value"
at this point is uninitialized kernel segment memory, and so dropping it
is not intrinsically safe. (Though it works for most types.)

This change switches to writing the container using core::ptr::write,
which is defined as moving its argument *without* dropping the contents
of the target.