tools: add usb/control-test
This rust program uses libusb on a USB host to perform a simple read and
write to a connected device.

As described in the program's documentation, it requires that the device
running Tock has loaded the app examples/tests/usb, which enables the
USB controller on the device and causes it to respond to control
requests.