Stick to a revision of compiler_builtins
The compiler-builtins crate provides compiler intrinsics for ARM.
However, it doesn't have any releases we could track and also seems to
ride on top of nightly. They recently updated to use associated
constants, which were released in a recently nightly but are not
available in the nightly currently used by Tock.

While we should definitely update Tock's nightly at some point, we
should be pegged to a certain revision of compiler-builtins to prevent
this kind of problem in the future.