Make alarm driver interface future compatible
Currently the alarm driver only supports a single alarm, and
virtualization happens in userland. This, however, makes the system call
interface future compatible for when we _do_ implement virtualization in
the kernel.

Also gets rid of a bunch of calls and logic that are no longer used for
doing timer logic in the kernel.