Increase write buffer length for nRF serialization capsule
The write buffer is only 256 Bytes in the nRF serialization capsule, but you can create characteristics that are up to 510 Bytes in length, which then cause buffer overruns in the kernel (which Rust does catch and panic about). Bumping it up to 600 like the read buffer fixes it.

Tested on a Signpost module.

No format fixes necessary.