Create design document for thread/6lowpan stack
This document will eventually serve as a complete interface guide for developers to use when writing code for the networking stack on Tock. It will describe the interfaces between layers, the nature of layering and encapsulation, and more. For now, the document is in its very early stages and likely contains some erroneous information in addition to lots of missing information.