doc: add networking stack design document
This document will eventually serve as a complete interface guide for
developers to use when writing code for the networking stack on Tock. It will
describe the interfaces between layers, the nature of layering and
encapsulation, and more. For now, the document is in its very early stages
and likely contains some erroneous information in addition to lots of missing
information.

Add
- a better stack architecture diagram
- an additional section that addresses file names
- revisions to the traits and implementation section
- Describe several message transmission paths and inter-layer dependencies