sam4l: flash: update driver
- Remove the "is flash actually erased check". This was a bit of a hack
since it required the flash to be ready without actually checking it
before issuing the quick page read. In addition, we weren't reading the
correct page and we weren't checking the return value, so this check has
been an NOP so far. We just remove it to keep things simpler.

- Otherwise simplify the driver. This get's rid of state variables that
were set but never checked and consolidates state variables.