More updates to Thread_Stack_Design.txt
These updates include a better stack architecture diagram, an additional
section that addresses file names, and revisions to the traits and
implementation section. This document is still incomplete and I suspect
it is not quite what we are looking for in several places, but I also
think it is now at a point where I would benefit from getting input from
others about the direction of the document so far.