Implemented Alarm to detect the "current app"
More or less copy-pasted the AlarmData struct from the alarm capsule
and it is a part of each process's grant

Every time an alarms fires the grants are searched to find the
process that expired or have the smallest value

However, in according to NRF Connect the advertisement interval don't match that
configured interval according to my phone but I will test is with my
logic analyzer

Added appid to the chip(radio.rs) which is used for example to determine which
app the callback belong to

Adapted nrf51/radio.rs to the new capsule but it don't work at moment.
Haven't debugged it but seem to hang somewhere suspect the new interrupt scheme.