Use more idiomatic pointer conversions instead of mem::transmute
mem::transmute is very unsafe as it does not really type check the
input value. The more idiomatic and more typesafe way is to use a safe
pointer conversion from usize to *const T followed by an unsafe
pointer-to-reference conversion via &*T. The advantage of using the
conversion over mem::transmute is that the compiler would complain as
soon as you stop using usize as an input value type.