Add explicit, optional calibration to CM4 SysTick
In some cases, the SysTick calibration is not set by the core. So, in
order to get 10s of ms units, the board/chip must tell us explicitly
what is the clock speed of the SysTick.

This change, first, creates a real struct for SysTick, similar to the
MPU struct, which contains a tenms field. If the core's tenms register
is 0, we use the struct's field instead. It also adds a constructor that
takes a clock speed and converts it to a tenms.

This change requires changing how SysTick is instantiated in chip crates
since constructors now return a struct instead of a reference, but it
should only be some small type changes, and it's more similar to the MPU
this way anyway.