Add MPU + SysTick to NRF52
Adds support for the MPU and SysTick to the NRF52 chip.

We use a calibration value of 64Mhz for the SysTick since the NRF52 does
not contain a callibration value in the core itself (and SysTick is
clocked by the CPU clock, which is always 64Mhz).

The MPU required fixing a "bug" in the NRF52DK's chip_layout, which was
setting the MIN_MPU_ALIGNMENT to 8 _bytes_ rather than 8K. Other boards
using the NRF52 will also have to ensure this value is set properly.

Finally, updates the SAM4L to use the very slighly modified SysTick
constructor interface.