Simplify BLE advertising timer implementation
This commig significantly simplifies how timers for BLE advertising and
scanning are handled. In the process it fixes some concurrency bugs
wherein the underlying alarm may be overriden incorrectly and adds the
pseudo random `advDelay` to the advertising interval required by the BLE
specification.

First, it separates the logic for determining the next alarm for each
app from determining which alarm to set with the underlying hardware.
One is app-specific (and therefore a method of the `App` struct), while
the latter is a global decision (and therefore a method of the `BLE`
struct).

Second, it gets rid of the circular buffer for keeping priorities of
apps. For now, it's just simpler to assume collisions can be dropped on
the floor. In any case, priorities are likely better implemented using a
per-app field that stores that last successful event, rather than an
explicit list.

These two let us get rid of a bunch of unnecessary code and meta-data.
It also better mirrors how the alarm capsule is implemented, which might
eventually lead to extracting some common code in the future.