Add the initial version I was working on.
I forked [@shaneleonard 's version of tock](shaneleonard/tock-teensy)
(where support for the teensy 3.6 was started),
to add support for the teensy 3.2, which I will use to
[build keyboard firmware](techhazard/dactyl-build).

I didn't like the layout of including tock itself as a submodule, so I
forked the [original project](helena-project/tock) and added my version
for the teensy 3.2 in the same way as the already supported boards.

I would very much like to thank both @shaneleonard and the people
behind the tock project for their work upon which I will build.

N.B.: This version is not yet correct at all!
Almost all of the values are currently from the 3.6 version

TODO:
- check the entire codebase (`boards/teensy3.2` and `chips/mk20`) and
  cross-reference [the mk20 reference manual](https://www.pjrc.com/teensy/K20P64M72SF1RM.pdf) with [the mk66 one](https://www.pjrc.com/teensy/K66P144M180SF5RMV2.pdf)
  to see where the two chipsets differ.
- find a good place for the macros in `common/`
- check if `libteensy` is necessary for anything besides `apps`
- compare original `Makefile`s and @shaneleonard 's, to make it compile again
- verify chip layout (`boards/teensy3.2/chip_layout.ld`)