update rustfmt (0.7.1 -> 0.3.4) *no code changes*
Yes, 0.7.1 -> 0.3.4 is an update, upstream reset versioning.

This commit only updates the tooling, the subsequent commit updates the
actual code, and minimal style changes such that the new formatter does
not exit with a hard error.

Branches should be able to cherry-pick just this commit and update
formatting locally to minimize merge churn.