kernel: regs: handle case when field is 32 bits
Because you can't shift past the size of the underlying type in rust, we
could not create the mask for a register field that is 32 bits using the
equation `(1<<numbits)-1`. This changes that calculation to be two
steps, which avoids the problem. It does however mean it won't work for
0, but we can't have a 0 bit register field.