sam4l: allow low-power operation of I2C peripheral
Critically, this allows the core to sleep while the PDCA clock (Peripheral DMA
Controller) is active, which according to p320:

    16.4.1 Power Management
      If the CPU enters a sleep mode that disables the PDCA clocks, the PDCA
      will stop functioning and resume operation after the system wakes up from
      sleep mode.

I think this is the behavior we actually want. If a peripheral has an
outstanding DMA transaction, it knows about it, and the _peripheral_ can
inhibit low-power mode. If the peripheral doesn't care, then let the core sleep.