kernel: process: check app ram length
Before we were trusting the app to request a reasonable amount of ram in
its header, and using that ram to set up pointers and other per-app
kernel-level state. However, if an app requests 0 ram we were still
setting up this state (causing a buffer overflow) and then panicing in
the MPU because we were trying to setup an MPU region with zero length.
This changes the app ram handling to ensure that the app gets at least
enough ram to setup the kernel state.

Apps can request 0 now, and they will likely still crash but it will be
an app fault and not a kernel fault.