allow access to FieldValue mask
I'm not sure if this is the right way to go or not. One approach would
be to keep adding more comparison methods to the registers object, in
this case something like `any_other_bits_set`, but I think that this is
a less common check.

I'm getting the value I want out of the registers system in the creation
of the mask. It's a semantically tricky thing to then "invert" the mask,
asking for all other fields that I didn't explicitly name already. At a
certain point, doing the bit manipulation direclty is clearer, and I
think this is that point.