sam4l: fix usart sleep during panics
Josh found a nice bug here. We special-cased panic for the usart manager,
but not quite enough. Specifically, panic doesn't use the usart peripheral
as the driver expects. When panic begins, it turns on the clock and
unconditionally expects the clock to stay on. If there's any outstanding
prints from debug, the debug `flush` will use the original usart driver to
complete these (critically, *after* the panic usart object has been created).
Thus when the `flush` print completes, the original driver (correctly) sees
that it has no more to do and disables usart clock out from underneath the
panic usart object.

The fix here is to just promote `IS_PANICING` to global state, and not tie
that state to a specific instance of usart manager.