userland: allow/enforce compile-time STACK_SIZE
Change the use of STACK_SIZE in crt1 to be an external variable.
We were already defining this as a symbol during linking, but we'd
also defined it at compile-time for libtock, so it was fixed.
This fixes to actually resolve at link time and further to protect
against libtock accidentally compiling with a fixed stack size in
the future.