travis: build course/sensys
The sensys course is what I point people to for getting started, and we
should ensure that it keeps compiling by adding it to travis. I don't
think it is too much of a burden to keep it up to date since it is
basically the same as hail.

fixes #765