Use slow start for Hail oscillator
Empirically, it seems as though fast start doesn't work on Hail after
some number of cycles into WAIT state (some number is in the dozens, but
basically running a blink app seems to often stall after maybe 100
delays). This _matches_ what was observed with Signpost a while ago and
was unexplained since Signpost and Hail use the same crystals.