cc26xx: correct the selection of LF clock source
This changes the clock source of the LF clock to use the RCOSC
derived clock instead of using the XOSC (external oscillator). When
switching to only running the LF clock (eg. during deep sleep) you need
to use the internally derived clock and not an clock with another power
source - this ensures that we default to the clock using the internal
power source (RCOSC) instead of the external.