Avoid concurrent PeripheralManager instances
Before this commit, the SAM4L's spi driver included nested calls to
functions that each acquire and release a SpiRegisterManager. The
acquisition isn't an issue, but the nested call might end up disabling
the clock before the caller is actually done using the registers.

This commit only creates a new SpiRegisterManager in "top-level"
functions and is careful never to call a nested function that also
acquires that lock.

Additionally, a small nit, it relies on the SPIEN bit in the status
register, rather than the `transfers_in_progress` counter to determine
when it's safe to disable the peripheral clock.