sam4l usb: add flow control for bulk transfers
This commit adds a flow-control protocol between the USB driver and its
client capsule, allowing interrupts to be turned off while waiting for
client code to resume sending or receiving data.

Two PC utilities were added in `tools/usb/` to exercise this feature.
They both take arbitrary data on stdin, push it out to the device,
and write to stdout whatever the device echoes back.

The tool `bulk-echo` does this in Rust with cross-platform features of
the libusb library, but is inefficient because it must use blocking
calls to do I/O.  The alternate tool `bulk-echo-fast` is written in C
and uses a single call poll() to service its stdin and libusb file
descriptors.  (For stdout, possibly-blocking I/O is used for
simplicity.)

Although the USB driver is not yet performance tuned (in particular it
does not use the hardware's double-buffering feature), this test
manages around 250kbps.