Compile processes to .tbf extension not .bin
Doesn't change anything about the format, but uses the .tbf extension to
make it clear the output binaries are TBFs rather than "binaries"
(whatever that means). This commit still retains the .bin (in addition
to the .tbf) in TAB files so existing versions of tockloader work, but
that would now be considered deprecated.

closes #904