Improve systick precision
Uses 64-bit arithmetic in the systick implementations to convert from
microseconds to tics. 64-bit division is a _very_ expensive operation in
Thumb, but as long as the microsecond value is a reasonably
large multiple of 10, the conversion is optimized by the compiler to
simple 32-bit division/multiplication.

To facilitate this optimization, the commit changes the systick
interface by replacing the `value` method with a `greater_than` method,
since we only ever use the value in order to do a comparison with a
relatively large multiple of 10 (500).

Closes #907