sam4l: usart: move tx done callback
This commit moves where the `transmit_complete` callback is fired from
when the DMA is done to when the USART done interrupt occurs. In the
USART state machine, the callback was being triggered when the USART was
in a "Completing TX" state, which means that any USART commands called
in the callback likely didn't work. By moving the callback to when we
actually declare the TX finished (and back to idle) the transmit
complete callback can be used for additional USART commands.

This was not an issue for us before because we were never using the
transmit complete callback to start new USART commands. All USART
actions were directed by an application syscall, which gave the USART HW
enough time to get to idle (or at least that is what I believe is
happening).

I've tested this by running the Hail app on Hail, and verifying that
both the console prints work and the BLE advertisements work.