`unwrap_or` -> `unwrap_or_else`
`unwrap_or_else` is lazily evaluated should be preferred over `unwrap_or` which
is eagerly evaluated. In this case it doesn't matter (probably) because it is
likely to inlined because it is an enum.

However, it more `idiomatic` to use `unwrap_or_else` and should preferred
especially if the something more `costly` such as calling another
function in the `else` case which will be evaluated always!