kernel: change exports
This goes through the lib.rs file in the kernel crate and tries to make
only things that need to be used outside of the kernel crate public.
This is better from a design perspective, but it should also prevent
problems in the future where `use` statements in other crates are not
consistent because things in the kernel crate were exported in two
different ways.