kernel: move main
main was in lib.rs, which was a little tricky to find. Also do_process
in sched.rs had to be public, but is only called in one place. This
merges the two, so all scheduling and main loop code is in the same
place.

It also renames the `main()` function to `kernel_loop()` to better
reflect what it actually is. `main()` implies the start of a program
with the normal argc/argv arguments, and this is neither.