sam4l: usart: spi: fix callback location
The SPI driver had the same issue as the UART driver: the callback was
triggered not when TX was in its Idle state, but instead when it was in
its Transfer Completing state. This means if you do anything in the
callback routine things will break.

This should also address the issue with the CS line going back high too
early, since now it is not set until it is safe to issue the callback,
which should definitely be at the end of the transaction.

Fixes #274