Use OpenOCD to flash Launchxl board
The existing makefile was using UniFlash, which is some TI-specific
tool. I couldn't quite get it to work, and OpenOCD (in git-tip) now has
support for the Launchxl's on-board debug chip (the xds110).

This commit adds an openocd script to flash the kernel hex, and changes
the Makefile to use it.