Assign proper gpio pins to nRF52-DK
In `nrf52dk_base` we had assumed some pins are the same but they were
not. Because for `nRF52DK` pin 19 and pin 20 were assigned to both the
LEDS and the SPI drivers!

However, now `setup_board` takes very many parameters which is not
efficient i.e, can't go in the registers but since this is during
the initilization I guess it is ok!