Add I2C master driver for CC26x2
Implements the I2C master driver for CC26x2. The I2C controller doesn't
support using the uDMA on the chip, so this driver _has_ to be operated
byte-by-byte.

Moreover, there is currently some overlap in the functionality
implemented in CC26x2 and CC26xx crates, specifically for power
management. When in doubt, I implemented logic in the CC26x2 version
since (a) I can't test on a CC2650, and (b) we should probably just
remove the CC26xx crate anyway.