kernel: add `kernel.create_grant()`
This is beneficial because:

1. It makes the API a little more consistent, and makes it clear that
board_kernel (the &kernel::Kernel) is the main access to the kernel API.
.kernel_loop() is also called from the Kernel struct.

2. It simplifies where the kernel API is defined. What operations does the
kernel provide? Look at the pub functions in the impl Kernel block. Note: this
is different than the interfaces and tools the kernel provides (like StaticRef,
Component, etc., and the HIL interfaces).

3. It would be straightforward to move the static variable CONTAINER_COUNTER to
the Kernel struct. Doing this would allow the kernel to verify that no grants
are created after processes are created (which would be invalid if any code
tried to do that because no space would be reserved for them in the process).