Fixed PRCM register offsets and HF clock
Just enough hacks to get a couple of the PRCM registers to align to
their expected offsets. there could be more offset issues, and we should
check!

Using the external HF oscillator doesn't seem to work reliably, so I
replaced it with the high frequency RC oscillator, which seems to work,
but who the heck knows what's actually going on