kernel: syscall: return why after switch_to
When a process stops executing and switches back to the kernel also
return why the process stopped. This removes the need for the scheduler
to ask why, and makes for a cleaner interface.

This also restructures the logic in `do_process` in sched.rs. It trades
off more indentation for an easier to follow flow.