kernel: remove TBF header v1
This has been deprecated for a while now, and there is very little
chance that any boards still have apps with the old tbf header version.

This also removes the PIC options field that is not supported by the
kernel, not standarized, and not something we want the kernel to have to
support in the future. It is also, as far as we know, unused.