Change MapCell to work with union restrictions
The problem is that a union used in a const _has_ to define all bytes of
the value, and those bytes have to at least be plausible---e.g., the tag
field of an enum can't be out of range.

Instead of creating uninitialized memory, we just store the union
directly in the MapCell, and pull out the `some` value lazily when we
know it's valid.