Use Cell instead of MapCell to store arch details
I don't have a great explanation for why MapCell causes a fault and Cell
doesn't, but my guess is that there is undefined behavior somewhere in
MapCell's implementation, and that is causing the compiler to do
unexpected things in unexpected places.

Switching to Cell fixes the bugs that manifest for now.