Fix assembly ordering in context switching code
Fixes a bug in the context switching code that appears under certain
circumstances where the compiler reorders volatile writes before inline
assembly (which was not marked volatile). The fix is to mark them
volatile as well as move the non-assembly code in hardfault_handler to a
separate function so we can make hardfault_handler naked like the other
ISRs.

This can lead to defer execution of system calls or handling of faults
which seems to interact in such a way that sometimes app code faults.

At a high level, the code to determine why a process was interrupted
(e.g. for a system call, timeslice, or fault) were moved to before
context switching to the app, so the kernel always thought the switch
was due to whatever the previous reason was.

This change also adds a new context switch condition, Interrupted, to
distinguish cases where the context switch was due to an ISR (and it is
used as the default case). I believe this is unrelated to the fix, but
is meaningful and was helpful when I was debugging.