boards: remove nrf51dk
This platform is not a good fit for Tock (it has a very simplistic MPU
and little RAM), and is not really used (a bug in its main.rs went
unnoticed for months before the 1.3 release). Given this, it is better
to move this out of mainline Tock to reduce the developer burden of
maintaining it.