boards/kernel_layout.ld: Remove `prog` section
We will add this section once we have user-space apps.

The reason for removing this is because from the linker, a symbol `_sapps` is
introduced, which is used in `boards/nucleo_f446re/src/main.rs`

We are not using `_sapps` at this time. We also need to `prog (rx)` section to
`boards/nucleo_f446re/chip_layout.ld` in order for linker to define `_sapps`.
Right now we are also not adding `prog (rx)` to
`boards/nucleo_f446re/chip_layout.ld`

Signed-off-by: Rajiv Ranganath <rajiv.ranganath@atihita.com>