libraries/tock-cells: Add `unwrap` function for `OptionalCell`
Introduce `unwrap` function with the following signature.

```
pub unsafe fn unwrap(&self) -> T
```

By specifying the function as `unsafe`, indicate to the caller that there are
extra invariants that the caller needs to consider as this method will panic if
the content is `None`.

With `unwrap` we can write

```
// Safety:
//   `xxx` should not be `None`
unsafe fn fun1() {
   let mut p = xxx.unwrap();
}
```

Instead of

```
// Safety:
//   `xxx` should not be `None`
unsafe fn fun1() {
   let mut p = xxx.expect("ABC");
}
```

And save on storage space for "ABC".

Signed-off-by: Rajiv Ranganath <rajiv.ranganath@atihita.com>