kernel/sched: Change `systick.enable` to `false`
`systick.enable` take a parameter of type `bool` to indicate if the SysTick
interrupt handler needs to be called or not.

While we need SysTick timer to be running before we enter the `do_process`
loop, from what I understand we do not require SysTick interrupt handler to be
enabled intially.

SysTick interrupt handler is enabled (using `systick.enable(true)`) prior to
switching to user space.

Signed-off-by: Rajiv Ranganath <rajiv.ranganath@atihita.com>