Asked Pat what this comment meant (since the registers are on the RF233 and accessed via SPI operations, you can't use the Register interface), he replied:
Honestly, I'm not sure what I was thinking there. Looking at the history, that looks like it was part of a bigger, somewhat mechanical, refactor. That was shortly after the new register interface was merged.. if I had to guess I suspect I had some notion that there would be some magical, "more elegant" means of handling constants coming, but I'll confess I'm not sure what that would be.

I think it should be fine to simply delete that comment. Sorry for the noise :/