Add initial RISC-V support
This commit is the first towards adding RISC-V support to Tock. It
includes enough code that a RISC-V core will boot and can handle
interrupts. It also includes basic support for the HiFive1 (a) board,
and some SiFive peripherals.

The HiFive1(a) board does not support User mode, so there is no support
for userland (yet). This simply allows kernel code to boot and run.

The SiFive peripherals are organized into their own folder because they
are shared among multiple SiFive cores.