cortex-m: update syscall to UKB v2
For portability reasons the syscall interface has been updated, and this
commit implements it for cortex-m.

The major change is that we no longer remove the SVC stack frame after a
yield call. In fact, the API for popping stack frames has been removed.
Instead, when we want the process to run a new function, we re-use the
old SVC stack frame. In the case that the app has never run before we do
have to create the stack frame just like we used to.

In theory, this is actually more efficient since we save the effort of
removing the stack frame, but that wasn't very substancial so who knows.

The other change is we now set state.psr and state.yield_pc after every
syscall, and not just effectively after yield(). I don't think this has
any effect.