Implement a driver, HIL, and capsule to write to flash
The purpose of the capsule is to provide low-level control of the embedded flash
to allow applications to implement flash-efficient data-structures using their
writeable flash regions. The API is blocking since most flash either halt the
CPU during write and erase operations or ask the application to wait until the
operation is finished. A blocking API is also simpler to reason and less
error-prone.