Add a nRF52-specific driver for the NVMC
The purpose of this driver is to provide low-level access to the embedded flash
of nRF52 boards to allow applications to implement flash-aware (like
wear-leveling) data-structures. The driver only permits applications to operate
on their writeable flash regions. The API is blocking since the CPU is halted
during write and erase operations.