boards: linker: add sbss sections
The RISC-V linker now creates a .sbss section which if not specified in
the linker file ends up as its own section, which gets placed in the
program binary (even though it should be in RAM). objdump places it at
the start of RAM, resulting in a very large .bin file.

Following the riscv-embedded Rust linker file, this commit adds the
sections in the same location as the .bss sections.

https://github.com/rust-embedded/riscv-rt/blob/2d0a00e6398ce82c528fef371d676c01429961c1/link.x#L69

I also recompiled Hail before and after and the resulting binary is the
same.