arty e21: remove PMP driver
For some reason, adding the PMP driver causes userspace to fail on the
E21 core running on the Arty FPGA. I don't know why, but reverting this
change seems to allow userspace to run.

The driver should be added back when someone is able to test on the E21.

Symptoms:
- App is loaded, and the kernel switches to the app.
- The first memop syscall is triggered, and execution switches back to
the kernel.
- When the kernel goes to resume the application a fault occurs. I
believe the `mcause` is set to 1 ("Instruction access fault").
- The app goes into a fault state and the kernel panics.