Implement the low-level-debug capsule.
The low-level-debug capsule provides app debugging functionality that can be accessed from inline assembly. Like the Console capsule, it provides a way for applications to report errors, but unlike the Console capsule it does not require the application to use the allow() or subscribe() system calls.

libtock-c and libtock-rs should be able to report low-level failures using this capsule when the app's state is unknown. If the capsule is not present then the diagnostics will be silently dropped.

This capsule follows the design proposed in issue #1302.