kernel: process: Return a slice if memory is already allocated
When calling the process alloc() function let's return a valid slice if
the new size is less then the current size. This means that even is we
don't need to allocate more memory (as we have enough) we still return
a usable slice.

This fixes running libtock-rs on the HiFive 1.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>